Move card, user to kaiten namespace, add comments
