+++ card_utils.py
@@ -1,10 +1,25 @@
 import requests
-from card import Card, CardType
+from kaiten.card import Card, CardType
 from kaiten.session import Session
 from typing import List, Iterable, Set
-from user import User
+from kaiten.user import User
 from dev_tasks import Dev_tasks
 
+def tag_from_card_type(type: CardType) -> str:
+    if type == CardType.Feature:
+        return ':F'
+    if type == CardType.User_story:
+        return ':US'
+    if type == CardType.Bug:
+        return ':BUG'
+    if type == CardType.Enabler:
+        return ':EN'
+
+""" type_id - тип карточек (сторя, энейблер...)
+    offset - сдвиг, если требуется запросить >100 карточек
+    limit - максимальное количество карточек для получения
+    query - текстовый запрос для поиска в названии карточки
+""" 
 def cards_type_request(session: Session, type_id: int, offset: int = 0, limit: int = 100, query: str = None):
     params = {
         "type_ids": type_id,
@@ -18,6 +33,10 @@ def cards_type_request(session: Session, type_id: int, offset: int = 0, limit: i
     return request.json()
 
 
+""" API Кайтена возвращает по 100 карточек максимум за раз. Для обхода ограничения 
+    используется поле 'offset' в запросе. С ним были какие-то проблемы, мы так и не выяснили.
+    Он мог пропускать какие-то карточки
+""" 
 def cards_of_type(session: Session, type_id: int, offset: int = 0) -> list:
     full_list = []
     temp_list = cards_type_request(session, type_id, offset)
@@ -27,16 +46,6 @@ def cards_of_type(session: Session, type_id: int, offset: int = 0) -> list:
         temp_list = cards_type_request(session, type_id, offset)
     full_list += temp_list
     return full_list
-
-def tag_from_card_type(type: CardType) -> str:
-    if type == CardType.Feature:
-        return ':F'
-    if type == CardType.User_story:
-        return ':US'
-    if type == CardType.Bug:
-        return ':BUG'
-    if type == CardType.Enabler:
-        return ':EN'
 
 
 def card_from_type(session: Session, type_id: CardType, identificator: str) -> Card | None:
@@ -59,7 +68,7 @@ def card_from_types(session: Session, type_ids: Set[CardType], identificator: st
 
     return None
 
-
+""" Отсортированный список карточек определенного типа, с подходящим ggis_id"""
 def cards_of_ggis_id(session: Session, type_id: CardType, id_tag: str) -> Iterable[Card]:
     card_list = []
     full_list = cards_of_type(session, type_id.value)
@@ -74,19 +83,19 @@ def cards_of_ggis_id(session: Session, type_id: CardType, id_tag: str) -> Iterab
 
 
 def features(session: Session) -> Iterable[Card]:
-    return cards_of_ggis_id(session, CardType.Feature, ':F')
+    return cards_of_ggis_id(session, CardType.Feature, tag_from_card_type(CardType.Feature))
 
 
 def user_stories(session: Session) -> Iterable[Card]:
-    return cards_of_ggis_id(session, CardType.User_story, ':US')
+    return cards_of_ggis_id(session, CardType.User_story, tag_from_card_type(CardType.User_story))
 
 
 def enablers(session: Session) -> Iterable[Card]:
-    return cards_of_ggis_id(session, CardType.Enabler, ':EN')
+    return cards_of_ggis_id(session, CardType.Enabler, tag_from_card_type(CardType.Enabler))
 
 
 def bugs(session: Session) -> Iterable[Card]:
-    return cards_of_ggis_id(session, CardType.Bug, ':BUG')
+    return cards_of_ggis_id(session, CardType.Bug, tag_from_card_type(CardType.Bug))
 
 
 def output_column(user: User, title: str) -> None:
