+++ dev_tasks.py
@@ -16,6 +16,9 @@ class Dev_tasks:
     def tasks(self) -> Iterable[str]:
         return iter(self.tasklist)
 
+""" Считать таски по формату: если первый символ строки не Tab, то считать это номером стори,
+    иначе - записать текст в список задач под последней попавшейся истории.
+"""
 def parse_tasks_file(path: str) -> List[Dev_tasks]:
     file1 = open(path, encoding='utf-8', mode='r')
     Lines = file1.readlines()
